k线页面快捷键功能修改

+++ app/src/main/java/com/jme/lsgoldtrade/domain/ContractInfoVo.java
@@ -175,7 +175,7 @@ public class ContractInfoVo implements Serializable {
         this.displayOrder = displayOrder;
     }
 
-    public int getBankLongMarginRate() {
+    public long getBankLongMarginRate() {
         return bankLongMarginRate;
     }
